requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
python-dotenv==1.0.1
pymongo[srv]==4.7.2
cloudinary==1.39.0
bcrypt==4.1.3
PyJWT==2.8.0
Flask-JWT-Extended==4.5.0
requests==2.32.3
werkzeug==2.3.7

